Ian's Individual Issue Fixed
This commit refactors all of the class errors into one error module so it is easier to keep track of where they all come from.
