+++ datasette/handle_exception.py
@@ -3,7 +3,7 @@ from .utils import add_cors_headers
 from .utils.asgi import (
     Base400,
 )
-from .views.base import DatasetteError
+from .views.error_module import DatasetteError
 from markupsafe import Markup
 import pdb
 import traceback
