+++ datasette/views/base.py
@@ -6,6 +6,7 @@ import textwrap
 import time
 import urllib
 from markupsafe import escape
+from .error_module import DatasetteError
 
 
 import pint
@@ -34,22 +35,6 @@ from datasette.utils.asgi import (
 
 ureg = pint.UnitRegistry()
 
-
-class DatasetteError(Exception):
-    def __init__(
-        self,
-        message,
-        title=None,
-        error_dict=None,
-        status=500,
-        template=None,
-        message_is_html=False,
-    ):
-        self.message = message
-        self.title = title
-        self.error_dict = error_dict or {}
-        self.status = status
-        self.message_is_html = message_is_html
 
 
 class View:
@@ -409,11 +394,6 @@ class DataView(BaseView):
         if self.ds.cors:
             add_cors_headers(response.headers)
         return response
-
-
-def _error(messages, status=400):
-    return Response.json({"ok": False, "errors": messages}, status=status)
-
 
 async def stream_csv(datasette, fetch_data, request, database):
     kwargs = {}
